requirements.txt
django-phonenumber-field==7.1.0
Django==4.2.5
phonenumberslite==8.12.44
django-environ==0.11.2
drf-spectacular==0.26.4
django-filter==23.2
django-ipware==5.0.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
